blackrock/requirements.txt
Bottleneck==1.1.0
cycler==0.10.0
lxml==3.6.1
matplotlib==1.5.1
numpy==1.11.1
pandas==0.18.1
Pillow==3.3.0
pyparsing==2.1.5
python-dateutil==2.5.3
python-pptx==0.5.8
pytz==2016.6.1
six==1.10.0
Unidecode==0.4.19
xlrd==1.0.0
XlsxWriter==0.9.3
